requirements.txt
requests==2.31.0
pandas==2.0.3
numpy==1.24.3
beautifulsoup4==4.12.2
psycopg2-binary==2.9.7
openpyxl==3.1.2
sqlalchemy==2.0.19
